+++ packages/blade/src/components/Avatar/Avatar.web.tsx
@@ -1,15 +1,21 @@
-import type { ReactElement } from 'react';
 import React from 'react';
 import type { AvatarProps } from './types';
 import { StyledAvatar } from './StyledAvatar';
 import { useAvatarGroupContext } from './AvatarGroupContext';
 import { AvatarButton } from './AvatarButton';
+import {
+  avatarToBottomAddonSize,
+  avatarToIndicatorSize,
+  avatarTopAddonOffsets,
+} from './avatarTokens';
 import { getStyledProps } from '~components/Box/styledProps';
 import { metaAttribute, MetaConstants } from '~utils/metaAttribute';
 import { assignWithoutSideEffects } from '~utils/assignWithoutSideEffects';
 import { throwBladeError } from '~utils/logger';
 import { UserIcon } from '~components/Icons';
 import type { BladeElementRef } from '~utils/types';
+import BaseBox from '~components/Box/BaseBox';
+import { getComponentId } from '~utils/isValidAllowedChildren';
 
 const getInitials = (name: string): string => {
   // Combine first and last name initials
@@ -31,6 +37,9 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
     href,
     target,
     rel,
+    isSelected,
+    bottomAddon: BottomAddon,
+    topAddon,
     // Image Props
     src,
     alt,
@@ -52,7 +61,7 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
     ...styledProps
   },
   ref,
-): ReactElement => {
+) => {
   if (__DEV__) {
     if (src && !alt && !name) {
       throwBladeError({
@@ -60,10 +69,17 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
         message: '"alt" or "name" prop is required when the "src" prop is provided.',
       });
     }
+    if (topAddon && getComponentId(topAddon) !== 'Indicator') {
+      throwBladeError({
+        moduleName: 'Avatar',
+        message: 'TopAddon only accepts `Indicator` component.',
+      });
+    }
   }
 
   const groupProps = useAvatarGroupContext();
   const avatarSize = groupProps?.size ?? size;
+  const isInteractive = Boolean(onClick || href);
 
   const commonButtonProps = {
     variant,
@@ -74,6 +90,7 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
     rel,
     onBlur,
     onFocus,
+    isSelected,
     onClick,
     onMouseLeave,
     onMouseMove,
@@ -112,6 +129,7 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
     return <AvatarButton {...commonButtonProps} icon={icon ?? UserIcon} />;
   };
 
+  const isSquare = variant === 'square';
   return (
     <StyledAvatar
       {...metaAttribute({ name: MetaConstants.Avatar, testID })}
@@ -119,8 +137,29 @@ const _Avatar: React.ForwardRefRenderFunction<BladeElementRef, AvatarProps> = (
       backgroundColor="surface.background.gray.intense"
       variant={variant}
       size={avatarSize}
+      isInteractive={isInteractive}
     >
-      {getChildrenToRender()}
+      <BaseBox width="100%" height="100%" position="relative">
+        {topAddon ? (
+          <BaseBox
+            position="absolute"
+            top={avatarTopAddonOffsets[variant][size].top}
+            right={avatarTopAddonOffsets[variant][size].right}
+          >
+            {React.cloneElement(topAddon, { size: avatarToIndicatorSize[size], display: 'block' })}
+          </BaseBox>
+        ) : null}
+        {getChildrenToRender()}
+        {BottomAddon ? (
+          <BaseBox
+            position="absolute"
+            bottom={isSquare ? '-10%' : '0%'}
+            right={isSquare ? '-10%' : '0%'}
+          >
+            <BottomAddon display="block" size={avatarToBottomAddonSize[size]} />
+          </BaseBox>
+        ) : null}
+      </BaseBox>
     </StyledAvatar>
   );
 };