+++ packages/blade/src/components/Table/index.ts
@@ -1,7 +1,26 @@
-export * from './Table';
-export * from './TableHeader';
-export * from './TableBody';
-export * from './TableFooter';
-export * from './TablePagination';
-export * from './TableToolbar';
-export * from './types';
+export { Table } from './Table';
+export { TableHeader, TableHeaderCell, TableHeaderRow } from './TableHeader';
+export { TableBody, TableCell, TableRow } from './TableBody';
+export { TableFooter, TableFooterCell, TableFooterRow } from './TableFooter';
+export { TablePagination } from './TablePagination';
+export { TableToolbar, TableToolbarActions } from './TableToolbar';
+export type {
+  TableProps,
+  TableBodyProps,
+  TableCellProps,
+  TableData,
+  TableEditableCellProps,
+  TableEditableDropdownCellProps,
+  TableFooterCellProps,
+  TableFooterProps,
+  TableFooterRowProps,
+  TableHeaderCellProps,
+  TableHeaderProps,
+  TableHeaderRowProps,
+  TableNode,
+  TablePaginationProps,
+  TableRowProps,
+  TableToolbarActionsProps,
+  TableToolbarProps,
+} from './types';
+export { TableEditableCell, TableEditableDropdownCell } from './TableEditableCell';